requirements.txt
-e .
aws-cdk.core==1.98.0
aws-cdk.core==1.98.0
aws-cdk.aws-dynamodb==1.98.0
aws-cdk.aws-apigateway==1.98.0
aws-cdk.aws-lambda==1.98.0
aws-cdk.aws-lambda-event-sources==1.98.0